+++ apps/mail/app/(full-width)/terms.tsx
@@ -1,6 +1,7 @@
 import { useCopyToClipboard } from '@/hooks/use-copy-to-clipboard';
 import { Card, CardHeader, CardTitle } from '@/components/ui/card';
 import { Github, ArrowLeft, Link2 } from 'lucide-react';
+import { Navigation } from '@/components/navigation';
 import { Button } from '@/components/ui/button';
 import Footer from '@/components/home/footer';
 import { createSectionId } from '@/lib/utils';
@@ -21,6 +22,7 @@ export default function TermsOfService() {
 
   return (
     <div className="relative flex min-h-screen w-full flex-col overflow-auto bg-white dark:bg-[#111111]">
+      <Navigation />
       <div className="relative z-10 flex flex-grow flex-col">
         {/* Back Button */}
         <div className="absolute left-4 top-4 md:left-8 md:top-8">
