+++ apps/mail/app/globals.css
@@ -96,6 +96,7 @@
   body {
     @apply bg-background text-foreground font-sans;
     overscroll-behavior: none;
+    @apply style-scrollbar;
   }
 }
