+++ apps/mail/components/home/HomeContent.tsx
@@ -29,36 +29,19 @@ import {
   ImageFile,
   Expand,
 } from '../icons/icons';
-import {
-  NavigationMenu,
-  NavigationMenuItem,
-  NavigationMenuLink,
-  NavigationMenuList,
-  NavigationMenuTrigger,
-  NavigationMenuContent,
-  ListItem,
-  navigationMenuTriggerStyle,
-} from '@/components/ui/navigation-menu';
 import { PixelatedBackground, PixelatedLeft, PixelatedRight } from '@/components/home/pixelated-bg';
-import { Sheet, SheetContent, SheetHeader, SheetTitle, SheetTrigger } from '@/components/ui/sheet';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
-import { Form, FormControl, FormField, FormItem } from '@/components/ui/form';
-import { use, useCallback, useEffect, useRef, useState } from 'react';
-import { signIn, useSession } from '@/lib/auth-client';
-import { useInView, motion } from 'motion/react';
-import { Link, useNavigate } from 'react-router';
 import { Button } from '@/components/ui/button';
 import { Balancer } from 'react-wrap-balancer';
-import { Input } from '@/components/ui/input';
-import { Command, Menu } from 'lucide-react';
-import { Separator } from '../ui/separator';
-import { useForm } from 'react-hook-form';
+import { signIn } from '@/lib/auth-client';
+import { Navigation } from '../navigation';
 import { useTheme } from 'next-themes';
-import { cn } from '@/lib/utils';
+import { use, useEffect } from 'react';
+import { motion } from 'motion/react';
+import { Link } from 'react-router';
 import { toast } from 'sonner';
 import Footer from './footer';
 import React from 'react';
-import { z } from 'zod';
 
 const firstRowQueries: string[] = [
   'Show recent design feedback',
@@ -77,96 +60,12 @@ const tabs = [
   { label: 'Write Better Emails', value: 'feature-3' },
 ];
 
-const resources = [
-  {
-    title: 'GitHub',
-    href: 'https://github.com/Mail-0/Zero',
-    description: 'Check out our open-source projects and contributions.',
-    platform: 'github' as const,
-  },
-  {
-    title: 'Twitter',
-    href: 'https://x.com/zerodotemail',
-    description: 'Follow us for the latest updates and announcements.',
-    platform: 'twitter' as const,
-  },
-  {
-    title: 'LinkedIn',
-    href: 'https://www.linkedin.com/company/zerodotemail/',
-    description: 'Connect with us professionally and stay updated.',
-    platform: 'linkedin' as const,
-  },
-  {
-    title: 'Discord',
-    href: 'https://discord.gg/0email',
-    description: 'Join our community and chat with the team.',
-    platform: 'discord' as const,
-  },
-];
-
-const aboutLinks = [
-  {
-    title: 'About',
-    href: '/about',
-    description: 'Learn more about Zero and our mission.',
-  },
-
-  {
-    title: 'Privacy',
-    href: '/privacy',
-    description: 'Read our privacy policy and data handling practices.',
-  },
-  {
-    title: 'Terms of Service',
-    href: '/terms',
-    description: 'Review our terms of service and usage guidelines.',
-  },
-];
-
-const betaSignupSchema = z.object({
-  email: z.string().email().min(9),
-});
-
 export default function HomeContent() {
-  const tabRefs = useRef<HTMLButtonElement[]>([]);
-  const [glowStyle, setGlowStyle] = useState({ left: 0, width: 0 });
-  const [open, setOpen] = useState(false);
   const { setTheme } = useTheme();
-  const ref = useRef(null);
-  const { data: session } = useSession();
-  const navigate = useNavigate();
 
   useEffect(() => {
     setTheme('dark');
   }, [setTheme]);
-
-  const handleTabChange = useCallback((value: string) => {
-    const index = tabs.findIndex((tab) => tab.value === value);
-    const tab = tabRefs.current[index];
-    if (tab) {
-      const rect = tab.getBoundingClientRect();
-      const listRect = tab.parentElement?.getBoundingClientRect();
-      const offsetLeft = listRect ? rect.left - listRect.left : 0;
-
-      const newWidth = rect.width * 0.9;
-      const newLeft = offsetLeft + (rect.width - newWidth) / 2;
-
-      setGlowStyle({ left: newLeft, width: newWidth });
-    }
-  }, []);
-
-  useEffect(() => {
-    if (tabs[0] && tabs[0].value) {
-      handleTabChange(tabs[0].value);
-    }
-  }, [handleTabChange]);
-
-  const IconComponent = {
-    github: GitHub,
-    twitter: Twitter,
-    discord: Discord,
-    linkedin: LinkedIn,
-  };
 
   return (
     <main className="relative flex h-full flex-1 flex-col overflow-x-hidden bg-[#0F0F0F]">
@@ -177,134 +76,8 @@ export default function HomeContent() {
           maskImage: 'linear-gradient(to bottom, black, transparent)',
         }}
       />
-      {/* Desktop Navigation - Hidden on mobile */}
-      <header className="fixed z-50 hidden w-full items-center justify-center px-4 pt-6 md:flex">
-        <nav className="border-input/50 flex w-full max-w-3xl items-center justify-between gap-2 rounded-xl border-t bg-[#1E1E1E] p-2 px-4">
-          <div className="flex items-center gap-6">
-            <Link to="/" className="relative bottom-1 cursor-pointer">
-              <img src="white-icon.svg" alt="Zero Email" width={22} height={22} />
-              <span className="text-muted-foreground absolute -right-[-0.5px] text-[10px]">
-                beta
-              </span>
-            </Link>
-            <NavigationMenu>
-              <NavigationMenuList className="gap-1">
-                <NavigationMenuItem>
-                  <NavigationMenuTrigger>Company</NavigationMenuTrigger>
-                  <NavigationMenuContent>
-                    <ul className="grid w-[400px] gap-3 p-4 md:w-[500px] md:grid-cols-1 lg:w-[600px]">
-                      {aboutLinks.map((link) => (
-                        <ListItem key={link.title} title={link.title} href={link.href}>
-                          {link.description}
-                        </ListItem>
-                      ))}
-                    </ul>
-                  </NavigationMenuContent>
-                </NavigationMenuItem>
-                <NavigationMenuItem>
-                  <NavigationMenuTrigger>Resources</NavigationMenuTrigger>
-                  <NavigationMenuContent>
-                    <ul className="grid w-[400px] gap-3 p-4 md:w-[500px] md:grid-cols-2 lg:w-[600px]">
-                      {resources.map((resource) => (
-                        <ListItem
-                          key={resource.title}
-                          title={resource.title}
-                          href={resource.href}
-                          platform={resource.platform}
-                        >
-                          {resource.description}
-                        </ListItem>
-                      ))}
-                    </ul>
-                  </NavigationMenuContent>
-                </NavigationMenuItem>
-                <NavigationMenuItem>
-                  <a href="/pricing">
-                    <Button variant="ghost" className="h-9">
-                      Pricing
-                    </Button>
-                  </a>
-                </NavigationMenuItem>
-              </NavigationMenuList>
-            </NavigationMenu>
-          </div>
-          <div className="flex gap-2">
-            <Button
-              className="h-8 bg-white text-black hover:bg-white hover:text-black"
-              onClick={() => {
-                if (session) {
-                  // User is logged in, redirect to inbox
-                  navigate('/mail/inbox');
-                } else {
-                  // User is not logged in, show sign-in dialog
-                  toast.promise(
-                    signIn.social({
-                      provider: 'google',
-                      callbackURL: `${window.location.origin}/mail`,
-                    }),
-                    {
-                      error: 'Login redirect failed',
-                    },
-                  );
-                }
-              }}
-            >
-              Sign in
-            </Button>
-          </div>
-        </nav>
-      </header>
 
-      {/* Mobile Navigation Sheet */}
-      <div className="md:hidden">
-        <Sheet open={open} onOpenChange={setOpen}>
-          <SheetTrigger asChild>
-            <Button variant="ghost" size="icon" className="fixed left-4 top-6 z-50">
-              <Menu className="h-6 w-6" />
-            </Button>
-          </SheetTrigger>
-          <SheetContent side="left" className="w-[300px] bg-[#111111] sm:w-[400px]">
-            <SheetHeader className="flex flex-row items-center justify-between">
-              <SheetTitle>
-                <img src="white-icon.svg" alt="Zero Email" width={22} height={22} />
-              </SheetTitle>
-              <a href="/login">
-                <Button className="w-full">Sign in</Button>
-              </a>
-            </SheetHeader>
-            <div className="mt-8 flex flex-col space-y-3">
-              <div className="space-y-3">
-                <Link to="/pricing" className="mt-2">
-                  Pricing
-                </Link>
-                {aboutLinks.map((link) => (
-                  <a key={link.title} href={link.href} className="block font-medium">
-                    {link.title}
-                  </a>
-                ))}
-              </div>
-              <a target="_blank" href="https://cal.com/team/0" className="font-medium">
-                Contact Us
-              </a>
-            </div>
-            <Separator className="mt-8" />
-            <div className="mt-8 flex flex-row items-center justify-center gap-4">
-              {resources.map((resource) => {
-                const Icon = IconComponent[resource.platform];
-                return (
-                  <Link
-                    key={resource.title}
-                    to={resource.href}
-                    className="flex items-center gap-2 font-medium"
-                  >
-                    {resource.platform && <Icon className="h-5 w-5" />}
-                  </Link>
-                );
-              })}
-            </div>
-          </SheetContent>
-        </Sheet>
-      </div>
+      <Navigation />
 
       <section className="z-10 mt-32 flex flex-col items-center px-4">
         <motion.div
@@ -313,24 +86,19 @@ export default function HomeContent() {
           transition={{ duration: 0.5 }}
           className="border-input/50 mb-6 inline-flex items-center gap-4 rounded-full border border-[#2A2A2A] bg-[#1E1E1E] px-4 py-1"
         >
-          <span className="flex items-center gap-2 text-sm">
-            <img
-              src="/yc-small.svg"
-              alt="Y Combinator"
-              className="rounded-[2px]"
-              width={18}
-              height={18}
-            />{' '}
-            Backed by Y Combinator
-          </span>
-          {/* <Link
-            href="https://x.com/zerodotemail"
-            target="_blank"
-            className="text-foreground flex items-center gap-1 rounded-full bg-gradient-to-b from-neutral-600 to-neutral-700 px-3 py-1 text-sm"
-          >
-            <span>Learn More</span>
-            <ArrowRight className="ml-1" />
-          </Link> */}
+          <Link to="https://yc.vc" target="_blank" className="flex items-center gap-2 text-sm">
+            Backed by
+            <span>
+              <img
+                src="/yc-small.svg"
+                alt="Y Combinator"
+                className="rounded-[2px]"
+                width={18}
+                height={18}
+              />
+            </span>
+            Combinator
+          </Link>
         </motion.div>
         <motion.h1
           initial={{ opacity: 0, y: 20 }}
@@ -387,38 +155,14 @@ export default function HomeContent() {
             className="mt-2"
           />
         </a>
-       
       </section>
 
       <section className="relative mt-10 hidden flex-col justify-center md:flex">
         <div className="bg-border absolute left-1/2 top-0 h-px w-full -translate-x-1/2 md:container xl:max-w-7xl" />
         <Tabs
           defaultValue="smart-categorization"
-          onValueChange={handleTabChange}
           className="flex w-full flex-col items-center gap-0"
         >
-          {/* <div className="tabs-container relative hidden w-full max-w-[40rem] justify-center md:flex md:max-w-max">
-            <TabsList className="relative h-fit w-full rounded-none bg-transparent pb-0 md:w-auto">
-              <div className="bg-border absolute -top-4 left-0 h-[calc(100%+16px)] w-px" />
-              <div className="bg-border absolute -top-4 right-0 h-[calc(100%+16px)] w-px" />
-
-              <CustomTabGlow glowStyle={glowStyle} />
-
-
-              {tabs.map((tab, index) => (
-                <TabsTrigger
-                  key={tab.value}
-                  value={tab.value}
-                  className="relative h-12 w-52 rounded-none !bg-transparent !shadow-none"
-                  ref={(el) => {
-                    if (el) tabRefs.current[index] = el;
-                  }}
-                >
-                  {tab.label}
-                </TabsTrigger>
-              ))}
-            </TabsList>
-          </div> */}
           <div
             className="relative bottom-2 flex w-full justify-center md:border-t"
             style={{ clipPath: 'inset(0 0 0 0)', height: '110%' }}
@@ -432,8 +176,6 @@ export default function HomeContent() {
                 className="absolute right-0 top-0 -z-10 hidden h-full w-auto translate-x-full opacity-50 md:block"
                 style={{ mixBlendMode: 'screen' }}
               />
-              {/* <div className="bg-border absolute -left-px -top-4 hidden h-4 w-px md:block" />
-              <div className="bg-border absolute -right-px -top-4 hidden h-4 w-px md:block" /> */}
               {tabs.map((tab) => (
                 <TabsContent key={tab.value} value={tab.value}>
                   <img
@@ -1596,7 +1338,6 @@ export default function HomeContent() {
     </main>
   );
 }
-
 const CustomTabGlow = ({ glowStyle }: { glowStyle: { left: number; width: number } }) => {
   return (
     <div
